image-encoder/requirements.txt
Pillow>=10.0.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
pytesseract>=0.3.10
easyocr>=1.7.0
pytest>=7.4.0
requests>=2.31.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-magic>=0.4.27; sys_platform != 'win32'
